% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_eqr.R
\name{calc_EQR}
\alias{calc_EQR}
\title{Calculate EQRs and WFD status classes from diatom metric and sample environmental data}
\usage{
calc_EQR(x, header)
}
\arguments{
\item{x}{An object of class \code{DIATOM_METRIC}, usually the output from function \code{\link{calc_Metric}}.}

\item{header}{Data frame containing sample and site environmental information for claculating the expected value of the metric.}
}
\value{
A object of class \code{DIATOM_EQR}, a list with the following named elements:
}
\description{
Calculate EQRs and WFD status classes from diatom metric and sample environmental data
}
\author{
Steve Juggins \email{Stephen.Juggins@ncl.ac.uk}
}
