% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_metric.R
\name{calc_Metric}
\alias{calc_Metric}
\title{Calculate water quality metrics from diatom data}
\usage{
calc_Metric(x, metric, dictionary = darleq3::darleq3_taxa, verbose = TRUE)
}
\arguments{
\item{x}{Data frame of diatom counts or relative abundance data}

\item{metric}{Diatom metric, one of "TDI3", "TDI4", "TDI5LM", "TDI5NGS", "LTDI1", "LTDI2", "DAM".}

\item{dictionary}{Diatom dictionary, a data frame with diatom taxon codes and indicator values for different metrics.  Defaults to the built-in DARLEQ3 dictionary.}
}
\value{
A object of class \code{DIATOM_METRIC}, a list with the following named elements:
}
\description{
Calculate water quality metrics from diatom data
}
\author{
Steve Juggins \email{Stephen.Juggins@ncl.ac.uk}
}
