% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{calc_Metrics}
\alias{calc_Metrics}
\title{Calculate diatom water quality metrics for metric \"families\"}
\usage{
calc_Metrics(x, metrics = "TDI5LM", verbose = TRUE)
}
\arguments{
\item{x}{Data frame of diatom count or relative abundance data}

\item{metric.type}{Metric family for calculation.  One of either "TDILM" for river LM TDI metrics (TDI3, TDI4 and TDI5LM), "TDINGS" for river NGS TDI metrics (TDI5NGS), "LTDILM" for lake TDI metrics (LTDI1 and LTDI2) or "DAMLM" for river diatom acidification metric.}
}
\value{
A matrix with 3 columns listing the number of taxa, the N2 effective number of taxa, and the maximum abundance for each sample (row) of \code{x}.
}
\description{
This is a wrapper function to \code{\link{calc_Metric}} that calculates all metrics in a given family.  Metric families and their constituent metrics are as follows: TDILM (TDI3, TDI4, TDI5LM - river TDI metrics for LM data); TDINGS (TDI5NGS - river TDI metric for NGS data); LTDILM (LTDI1, LTDI2 - lake TDI metrics for LM data); DAMLM (DAM - river acidification metric for LM data).  For example, choosing metric family "TDILM" will calculate metrics TDI3, TDI4 and TDI5LM.
}
\references{
Kelly, M., S. Juggins, R. Guthrie, S. Pritchard, J. Jamieson, B. Rippey, H. Hirst, and M. Yallop, Assessment of ecological status in UK rivers using diatoms. \emph{Freshwater Biology}, 2008. 403-422.

Juggins, S., M. Kelly, T. Allott, M. Kelly-Quinn, and D. Monteith, A Water Framework Directive-compatible metric for assessing acidification in UK and Irish rivers using diatoms. \emph{Science of The Total Environment}, 2016. 671-678.

Bennion, H., M.G. Kelly, S. Juggins, M.L. Yallop, A. Burgess, J. Jamieson, and J. Krokowski, Assessment of ecological status in UK lakes using benthic diatoms. \emph{Freshwater Science}, 2014. 639-654.
}
\author{
Steve Juggins \email{Stephen.Juggins@ncl.ac.uk}
}
