% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{calc_N_N2_Max}
\alias{calc_N_N2_Max}
\title{Calculate summaries for diatom samples}
\usage{
calc_N_N2_Max(x)
}
\arguments{
\item{x}{data frame of diatom count or relative abundance data}
}
\value{
A matrix with 3 columns listing the number of taxa (N), Hill's N2, the effective number of taxa, and the maximum abundance (Max) of a taxon in each sample (row) of \code{x}.
}
\description{
Calculate summaries for diatom samples
}
\note{
This is a helper function and is not meant to be called by the user directly.
}
\references{
Hill, M.O. (1973). Diversity and evenness: a unifying notation and its consequences. \emph{Ecology} 54, 427--473.
}
\author{
Steve Juggins \email{Stephen.Juggins@ncl.ac.uk}
}
