% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{calc_SiteEQR}
\alias{calc_SiteEQR}
\title{Calculate site EQRs}
\usage{
calc_SiteEQR(EQR, SiteID, lake_Type = NULL)
}
\arguments{
\item{EQR}{sample EQR}

\item{SiteID}{site ID}

\item{lake_Type}{lake type following GB lake typology}
}
\description{
Calculate site EQRs
}
\details{
This is an internal function and is not meant to be called directly.
}
