% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{calc_Uncertainty}
\alias{calc_Uncertainty}
\title{Calculate classification uncertainties}
\usage{
calc_Uncertainty(x, metric, lake_Type = NULL)
}
\arguments{
\item{x}{data frame outfdrom from function \code{\link{calc_SiteEQR}}}

\item{metric}{diatom metric}

\item{lake_Type}{lake type following GB lake typology}
}
\description{
Calculate classification uncertainties
}
\details{
This is an internal function and is not meant to be called directly.
}
