% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{calc_WFDClass}
\alias{calc_WFDClass}
\title{Calculate WFD quality class}
\usage{
calc_WFDClass(EQR, metric, lake_Type = NA)
}
\arguments{
\item{EQR}{sample EQR}

\item{metric}{metric to calculate}

\item{lake_Type}{lake type following GB lake typology}
}
\description{
Calculate WFD quality class
}
\details{
This is an internal function and is not meant to be called directly.
}

