% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darleq3_fun.R
\name{darleq}
\alias{darleq}
\title{Calculate diatom water quality metrics, EQRs and WFD class for samples, and uncertainties of site classification}
\usage{
darleq(inFile, sheet = NULL, metrics = c("TDI3", "TDI4", "TDI5LM"),
  outFile = NULL, verbose = TRUE)
}
\arguments{
\item{inFile}{Excel file name containing diatom and sample environmental data. See \code{\link{read_DARLEQ}} for acceptible formats for these data.}

\item{sheet}{name of the worksheet in the Excel file to import.}

\item{metrics}{character vector of metric codes.  Currently one or more of the following: "TDI3", "TDI4", "TDI5LM" (for river LM TDI calculations), "TDI5NGS" for river NGS metric, "LTDI1", "LTDI2" for lake LM TDI metric or "DAMLM" for river diatom acidification metric.}

\item{outFile}{name of Excel file to save results.  If not given the function will generate a name by concatenating "DARLEQ3_Results_" with the original filename, the sheet name and the current date.}

\item{verbose}{logical to indicate should function stop immediately on error (TRUE) or return a \code{simpleError} (FALSE).  Defaults to TRUE.}
}
\description{
Calculate diatom water quality metrics, EQRs and WFD class for samples, and uncertainties of site classification
}
\details{
This is a wrapper function to \code{\link{read_DARLEQ}} \code{\link{calc_Metric_EQR}} and \code{\link{save_DARLEQ}} that imports data form an Excel file, calculates multiple metrics, EQRs and WFD classes and saves the results to another Excel file in one step.
}
\examples{
fn <- system.file("example_datasets/DARLEQ2TestData.xlsx", package="darleq3")
darleq(fn, outFile="Results.xlsx")

}
\references{
Kelly, M., S. Juggins, R. Guthrie, S. Pritchard, J. Jamieson, B. Rippey, H. Hirst, and M. Yallop, Assessment of ecological status in UK rivers using diatoms. \emph{Freshwater Biology}, 2008. 403-422.

Juggins, S., M. Kelly, T. Allott, M. Kelly-Quinn, and D. Monteith, A Water Framework Directive-compatible metric for assessing acidification in UK and Irish rivers using diatoms. \emph{Science of The Total Environment}, 2016. 671-678.

Bennion, H., M.G. Kelly, S. Juggins, M.L. Yallop, A. Burgess, J. Jamieson, and J. Krokowski, Assessment of ecological status in UK lakes using benthic diatoms. \emph{Freshwater Science}, 2014. 639-654.
}
\author{
Steve Juggins \email{Stephen.Juggins@ncl.ac.uk}
}
