% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darleq3.R
\docType{package}
\name{darleq3}
\alias{darleq3}
\alias{darleq3-package}
\title{darleq3: Diatom Assessment of River and Lake Ecological Quality (Version 0.7.1)}
\description{
darleq3 is an R package for the assessment of river and lake ecological status using diatom data obtained by light microscopy (LM) or Next Generation Sequencing (NGS). The package contains functions for importing data from Excel worksheets, calculating various water quality metrics, EQRs and Water Framework Directive quality classes.

Additonal help on using the package and interpreting the results can be found in the following vignettes:

\href{../doc/darleq3UserGuide.html}{darleq3 User Guide (html version)}

\href{../doc/darleq3UserGuidePDF.pdf}{darleq3 User Guide (pdf version)}

\href{../doc/darleq3.pdf}{darleq3 Manual (pdf version)}

\href{../doc/interpreting_NGS_data.pdf}{Guide to interpreting TDI5 NGS data (pdf version)}
}
\section{darleq3 functions}{

\itemize{
\item \code{\link{darleq}} import diatom data from an Excel file, calculate matrics, EQRs and WFD quality classes, and save results in Excel format
\item \code{\link{read_DARLEQ}} import diatom data from an Excel file
\item \code{\link{save_DARLEQ}} save metric and EQR results in an Excel file
\item \code{\link{calc_Metric}} calculate various diatom water quality metrics
\item \code{\link{calc_EQR}} calculate sample and site EQRs and WFD quality classes
\item \code{\link{calc_Metric_EQR}} calculate EQRS, WFD quality classes and summary diagnostic measures for multiple metrics
\item \code{\link{runDARLEQ}} run DARLEQ3 as an interactive shiny app in a web browser
}
}

\section{Acknowledgements}{

DARLEQ was developed and funded as part of the following projects:
\itemize{
\item Diatoms as Monitors of Ecological Status of Rivers (Diatom Assessment of River Ecological Status - DARES) Project No: EMC/WP04/078 Funded by the Environment Agency and SNIFFER (The Scotland and Northern Ireland Forum for Environmental Research).
\item Development of a phytobenthos classification tool for lakes and lochs of UK (DALES - Diatom assessment of lake and loch ecological status) Project Code: EMC/WP09/079. Funded by The Environment Agency Science Programme.
\item Development of the Diatom classification tool (DARLEQ) for lakes and rivers. Funded by The Environment Agency Science Programme.
}
}

\references{
Kelly, M., S. Juggins, R. Guthrie, S. Pritchard, J. Jamieson, B. Rippey, H. Hirst, and M. Yallop, Assessment of ecological status in UK rivers using diatoms. \emph{Freshwater Biology}, 2008. 403-422.

Juggins, S., M. Kelly, T. Allott, M. Kelly-Quinn, and D. Monteith, A Water Framework Directive-compatible metric for assessing acidification in UK and Irish rivers using diatoms. \emph{Science of The Total Environment}, 2016. 671-678.

Bennion, H., M.G. Kelly, S. Juggins, M.L. Yallop, A. Burgess, J. Jamieson, and J. Krokowski, Assessment of ecological status in UK lakes using benthic diatoms. \emph{Freshwater Science}, 2014. 639-654.
}
