% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{darleq3_data}
\alias{darleq3_data}
\title{darleq3_data internal variables used in metric, EQR and WFD quality class calculations}
\description{
darleq3_data internal variables used in metric, EQR and WFD quality class calculations
}
