% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{darleq3_taxa}
\alias{darleq3_taxa}
\title{darleq3_taxa DARLEQ3 diatom dictionary}
\description{
A dataframe containing diatom codes, names and indicator values for different water quality metrics
}
