% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{errMessage}
\alias{errMessage}
\title{error message handler}
\usage{
errMessage(txt, verbose)
}
\arguments{
\item{txt}{error message}

\item{verbose}{logical to indicate should function stop immediately on error (TRUE) or return a \code{simpleError} (FALSE).  Defaults to TRUE.}
}
\description{
error message handler
}
\details{
This is an internal function and is not meant to be called directly.
}
