% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darleq_obs.R
\name{print.DARLEQ_DATA}
\alias{print.DARLEQ_DATA}
\title{Print an object of class DARLEQ_Data}
\usage{
\method{print}{DARLEQ_DATA}(x, ...)
}
\arguments{
\item{x}{object of class DARLEQ_DATA}

\item{...}{additional arguments to \code{print}}
}
\description{
Print an object of class DARLEQ_Data
}
