% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_darleq.R
\name{read_DARLEQ}
\alias{read_DARLEQ}
\title{Read diatom data from an Excel file in DARLEQ2 format}
\usage{
read_DARLEQ(filename, sheet = NULL, verbose = TRUE)
}
\arguments{
\item{filename}{NAme of Excel file.}

\item{sheet}{Name of sheet within Excel file.}
}
\value{
A list with the following named elements:
}
\description{
Read diatom data from an Excel file in DARLEQ2 format
}
\author{
Steve Juggins \email{Stephen.Juggins@ncl.ac.uk}
}
