% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_darleq.R
\name{save_DARLEQ}
\alias{save_DARLEQ}
\title{Saves results of diatom EQR and class claculation to an Excel file}
\usage{
save_DARLEQ(d, outFile = NULL, fn = "", sheet = "", verbose = TRUE)
}
\arguments{
\item{d}{list of sample and site EQR WFD class results, usually the output from \code{\link{calc_Metric_EQR}}}

\item{outFile}{name of Excel file to save results.}

\item{fn}{name of in the input file for inclusion in the job summary page of output.}

\item{sheet}{name of the input worksheet for inclusion in the job summary page of output.}

\item{verbose}{logical to indicate if function should stop immediately on error (TRUE) or return a \code{simpleError} (FALSE).  Defaults to TRUE.}
}
\description{
Saves results of diatom EQR and class claculation to an Excel file
}
\examples{
fn <- system.file("extdata/DARLEQ2TestData.xlsx", package="darleq3")
d <- read_DARLEQ(fn, "Rivers TDI Test Data")
x <- calc_Metric_EQR(d)
save_DARLEQ(x, outFile="results.xlsx")

}
\author{
Steve Juggins \email{Stephen.Juggins@ncl.ac.uk}
}
