% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_eqr.R
\name{calc_EQR}
\alias{calc_EQR}
\title{Calculate EQRs and WFD status classes from diatom metric and sample environmental data}
\usage{
calc_EQR(x, header, verbose = TRUE)
}
\arguments{
\item{x}{an object of class \code{DIATOM_METRIC}, usually the output from function \code{\link{calc_Metric}}.}

\item{header}{data frame containing sample and site environmental information for calculating the expected value of the metric.}

\item{verbose}{logical to indicate should function stop immediately on error (TRUE) or return a \code{simpleError} (FALSE).  Defaults to TRUE.}
}
\value{
A object of class \code{DIATOM_EQR}, a list with the following named elements:
\item{EQR}{data frame containing, for each sample, sample codes, water chemistry data and other columns from the header in original Excel file, metric and summary information from function \code{calc_Metric}, expected EQRs (eEQR), calculated EQRs, predicted WFD class, percentage diatoms in diagnostic ecological groups, and a flag to indicate missing or out of range environmental data.}
\item{Uncertainty}{data frame containing, for each site, mean EQRS, predicted WFD class, and confidence of class (CoC) for each WFD class and HG/MPB boundary (CoCCHG, COCMPB), and risk of misclassification for the predicted class (ROM) and for the G/M boundary (ROM_GM)}.
}
\description{
Calculate EQRs and WFD status classes from diatom metric and sample environmental data
}
\examples{
fn <- system.file("extdata/DARLEQ2TestData.xlsx", package="darleq3")
d <- read_DARLEQ(fn, "Rivers TDI Test Data")
x <- calc_Metric(d$diatom_data, metric="TDI4")
eqr <- calc_EQR(x, d$header)
head(eqr$EQR)
head(eqr$Uncertainty)

}
\references{
Kelly, M., S. Juggins, R. Guthrie, S. Pritchard, J. Jamieson, B. Rippey, H. Hirst, and M. Yallop, Assessment of ecological status in UK rivers using diatoms. \emph{Freshwater Biology}, 2008. 403-422.

Juggins, S., M. Kelly, T. Allott, M. Kelly-Quinn, and D. Monteith, A Water Framework Directive-compatible metric for assessing acidification in UK and Irish rivers using diatoms. \emph{Science of The Total Environment}, 2016. 671-678.

Bennion, H., M.G. Kelly, S. Juggins, M.L. Yallop, A. Burgess, J. Jamieson, and J. Krokowski, Assessment of ecological status in UK lakes using benthic diatoms. \emph{Freshwater Science}, 2014. 639-654.
}
\author{
Steve Juggins \email{Stephen.Juggins@ncl.ac.uk}
}
